Update Master_Defense_O2.gsm
I added two selectors ('if' instances) to the functions HeDAR_0 and HeDAR2_0. These functions will have negative and values smaller than 1 if eDAR<eDAR_min=0.1 or values higher than 1 or 2 if eDAR>eDAR_max=1.5. To prevent this to happen, which would violate the logic that modified Hill functions should still range from 0 to 1 or 1 to 2, the selectors force HeDAR and HeDAR_2 to be 0 and 1 (or 1 and 2) for values smaller and higher than 0.1 and1.5, respectively.
The growth rates have been modified to account for these selectors.